Replace `TRUE/FALSE` with `IS_TRUE/IS_FALSE` on macros
This fix projects/compilers that has defined `TRUE` as a macro.
As an example, using cpp_mock with mingw will result on mocked method names
not being identified as defined types when the number of parameters where one or zero.
That was caused because `DEFINE_EXISTS` macro used `TRUE` wich was
expanded to `1`, and that being concatenated to `MOCK_METHOD_SINGLE_`.
Because `MOCK_METHOD_SINGLE_1` was not defined as a macro the preprocesed
code was invalid.